+++ packages/contracts/test/dollar/core/UbiquityDollarToken.t.sol
@@ -90,8 +90,8 @@ contract UbiquityDollarTokenTest is LocalTestHelper {
     }
 
     function testUUPS_InitializedVersion() external {
-        uint expectedVersion = 1;
-        uint baseExpectedVersion = 255;
+        uint256 expectedVersion = 1;
+        uint256 baseExpectedVersion = 255;
 
         UbiquityDollarTokenUpgraded newImpl = new UbiquityDollarTokenUpgraded();
         UbiquityDollarTokenUpgraded newImplT = new UbiquityDollarTokenUpgraded();
